requirements.txt
langchain
langchain_groq
langchain_core
langchain_community
python-dotenv
streamlit